requirements.txt
google-generativeai
anyio
trio
mcp
python-dotenv
openai
websockets
